+++ Server/docker-compose.yml
@@ -2,13 +2,11 @@ version: '3'
 
 services: 
   api:
-    build:
-      context: .
-      dockerfile: ./api/Dockerfile
+    build: .
     ports: 
       - '8080:8080'
     volumes: 
-      - './api:/go/src/api'
+      - '.:/go/src'
     tty: true
     stdin_open: true
   db:
