+++ Server/go.mod
@@ -1,0 +1,9 @@
+module api
+
+go 1.15
+
+require (
+	github.com/labstack/echo v3.3.10+incompatible
+	github.com/labstack/gommon v0.3.0 // indirect
+	golang.org/x/crypto v0.0.0-20201221181555-eec23a3978ad // indirect
+)
